Move full form to EditPost component
